requirements.txt
asgiref==3.7.2
attrs==23.2.0
boto3==1.20.26
botocore==1.23.54
cachetools==4.2.4
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.7
django-anymail==10.2
django-cors-headers==3.14.0
django-environ==0.11.2
django-import-export==4.0.7
django-jazzmin==2.6.0
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
environs==10.0.0
facebook-sdk==3.1.0
Faker==25.5.0
future==1.0.0
google-auth==1.21.2
google-auth-httplib2==0.0.4
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
marshmallow==3.20.1
oauthlib==3.1.0
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-twitter==3.5
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.0
rpds-py==0.18.1
rsa==4.9
s3transfer==0.5.2
setuptools==70.0.0
shortuuid==1.0.11
six==1.16.0
sqlparse==0.4.4
stripe==7.9.0
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.18
whitenoise==6.6.0
